Separated out the Splash viz from Main SplashViz is literally untestable in junit. Only causes a decrease in covered code.
