+++ codecov.yml
@@ -20,6 +20,8 @@ coverage:
     - .*/org/tros/torgo/swing/JConsole.*
     - .*/org/tros/torgo/viz/SpringUtilities.*
     - .*/org/tros/utils/swing/JLinkButton.*
+#literally untestable in junit
+    - .*/org/tros/torgo/MainSplash.*
 parsers:
   gcov:
     branch_detection:
