+++ licenseheader.txt
@@ -1,7 +1,7 @@
 <#if licenseFirst??>
 ${licenseFirst}
 </#if>
-${licensePrefix}Copyright 2015-2016 Matthew Aguirre
+${licensePrefix}Copyright 2015-2017 Matthew Aguirre
 ${licensePrefix}
 ${licensePrefix}Licensed under the Apache License, Version 2.0 (the "License");
 ${licensePrefix}you may not use this file except in compliance with the License.
